+++ TEAMSCODE/2023-SPRING/E.cpp
@@ -1,0 +1,18 @@
+#include <bits/stdc++.h>
+using namespace std;
+int main(){
+    int T; cin >> T;
+    while(T--){
+        int n, k; cin >> n >> k;
+        vector<int> a(n), b(n);
+        for(int i = 0; i < n; i++){
+            cin >> a[i];
+        }
+        for(int i = 0; i < n; i++){
+            cin >> b[i];
+        }
+        for(int i = n-1; i > n-k; i--){
+            
+        }
+    }
+}